main-view
Added a main-view.jsx, where movie/genre/director informatino as added as well as importing the movie-card/movie-view that was being built as well. added an on event click for both clicking on the movie to bring up the information and clicking on the back button to take you back to the list of movies.