updates
Updated mainview to remove the search functions that were built as they werent' working. also updated the Movieview (106) for work on the profile page to view favorited movie.

Movie-View updated to allow the adding/removal of favorite movies to the user signed in.

profilte view-updated to allow the users favorited movie card to be displayed on their page and linked to the full page.